requirements.txt
colorama
openai
python-dotenv
